+++ addon/objects/uml-primitives/fd-uml-class.js
@@ -315,11 +315,19 @@ joint.shapes.flexberry.uml.ClassView = joint.shapes.flexberry.uml.PrimitiveEleme
     }.bind(this));
 
     this.$box.find('.class-stereotype-input').on('blur', function(evt) {
-      this.showNormalizedStereotypeOnInput($(evt.target));
+      let stereotypeText = $(evt.target).val();
+      let stereotype = this.normalizeStereotype(stereotypeText);
+      let rows = stereotypeText.split(/[\n\r|\r|\n]/);
+      let $stereotypeInput = this.$box.find('.class-stereotype-input');
+      $stereotypeInput.val(stereotype);
+      $stereotypeInput.prop('rows', rows.length);
+      let objectModel = this.model.get('objectModel');
+      objectModel.set('stereotype', stereotype);
+      this.paper.trigger('updaterepobj', objectModel, 'stereotype', stereotype);
+      this.updateRectangles();
     }.bind(this));
 
     this.updateInputValue();
-    this.showNormalizedStereotypeOnInput(this.$box.find('.class-stereotype-input'));
 
     // Update the box position whenever the underlying model changes.
     this.model.on('change', this.updateBox, this);
@@ -429,7 +437,7 @@ joint.shapes.flexberry.uml.ClassView = joint.shapes.flexberry.uml.PrimitiveEleme
     classNameInput.prop('rows', objectModel.get('name').split(/[\n\r|\r|\n]/).length || 1);
     classNameInput.val(objectModel.get('name'));
     classStereotypeInput.prop('rows', objectModel.get('stereotype').split(/[\n\r|\r|\n]/).length || 1);
-    classStereotypeInput.val(objectModel.get('stereotype'));
+    classStereotypeInput.val(this.normalizeStereotype(objectModel.get('stereotype')));
 
     attributesInput.prop('rows', objectModel.get('attributes').length || 1);
     attributesInput.val(objectModel.get('attributes').join('\n'));
@@ -508,14 +516,5 @@ joint.shapes.flexberry.uml.ClassView = joint.shapes.flexberry.uml.PrimitiveEleme
     }
 
     return stereotype;
-  },
-
-  showNormalizedStereotypeOnInput(element) {
-    let stereotypeText = element.val();
-    let stereotype = this.normalizeStereotype(stereotypeText);
-    let rows = stereotypeText.split(/[\n\r|\r|\n]/);
-    element.val(stereotype);
-    element.prop('rows', rows.length);
-    this.updateRectangles();
   }
 });